Moved the navigation methods to screen actions, added deleting hubs in the cache datasource and removed the Id field from hubs and items bottom sheet
