+++ hub/core-hub/src/main/java/com/trackhub/core_hub/data/cache/dao/HubDao.kt
@@ -33,6 +33,9 @@ interface HubDao {
     @Query("SELECT * FROM hubs WHERE is_owned = 0")
     fun getSharedHubs(): Flow<List<HubEntity>>
 
+    @Query("DELETE FROM hubs")
+    suspend fun deleteAllHubs()
+
     @Upsert
     suspend fun updateHubItems(items: List<HubItemEntity>)
