+++ network/feat-network/src/main/java/com/trackhub/feat_network/domain/cache/CacheDataSource.kt
@@ -12,6 +12,7 @@ interface CacheDataSource {
     suspend fun updateSharedHubs(hubs: List<Hub>)
     fun getOwnHubs(): Flow<List<Hub>>
     fun getSharedHubs(): Flow<List<Hub>>
+    suspend fun deleteAllHubs()
     suspend fun updateHubItems(items: List<HubItem>)
     suspend fun deleteItems(items: List<HubItem>)
     fun getItemsFromHub(hubId: String): Flow<List<HubItem>>
